Refactor total analytics into a separate component
